renaming metacat-type-converter package
