+++ metacat-schema-converters/src/main/java/com/netflix/metacat/shema/converters/SchemaConverterModule.java
@@ -3,7 +3,7 @@ package com.netflix.metacat.shema.converters;
 import com.google.inject.AbstractModule;
 
 /**
- * Created by zhenl on 1/7/17.
+ * Schema converter module.
  */
 public class SchemaConverterModule extends AbstractModule {
     @Override
