+++ metacat-type-converters/src/main/java/com/netflix/metacat/types/Base.java
@@ -11,7 +11,7 @@
  *    limitations under the License.
  */
 
-package com.netflix.metacat.canonical.types;
+package com.netflix.metacat.types;
 
 /**
  Canonical base type class.
