+++ metacat-type-converters/src/test/groovy/com/netflix/metacat/type/converters/PigTypeConverterSpec.groovy
@@ -1,7 +1,7 @@
 package com.netflix.metacat.type.converters
 
-import com.netflix.metacat.canonical.types.TypeManager
-import com.netflix.metacat.canonical.types.TypeRegistry
+import com.netflix.metacat.types.TypeManager
+import com.netflix.metacat.types.TypeRegistry
 import spock.lang.Shared
 import spock.lang.Specification
